move fetch/pull impl to GitRemotoe
Created using spr 1.3.8-beta.1
